% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureEx.r
\name{featureEx}
\alias{featureEx}
\title{Feature Exploration}
\usage{
featureEx(x)
}
\arguments{
\item{x}{A single data.table or a list of data.tables of features.}
}
\value{
A list of various exploratory plots.
}
\description{
Feature Exploration
}
\examples{
x <- runif(100)
y <- x + rnorm(100)
z <- x + max(y)
toyData <- cbind(x,y,z)

}

