% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmpp.R
\name{kmpp}
\alias{kmpp}
\title{K-means++ initialization followed by K-means.}
\usage{
kmpp(x, k = 2)
}
\arguments{
\item{x}{A matrix, data.frame, or data.table}

\item{k}{An integer denoting how many centers.}
}
\value{
A kmeans object.
}
\description{
K-means++ initialization followed by K-means.
}
\examples{
data(iris)
kmpp(iris[,1:4],k=3)
}

