% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primitiveEx.r
\name{primitiveEx}
\alias{primitiveEx}
\title{Primitive Exploration}
\usage{
primitiveEx(x, coi)
}
\arguments{
\item{x}{A list of data.frames or data.tables of primitives}
}
\value{
A list of various plots using dimension reduction.
}
\description{
Primitive Exploration
}
\details{
Details to come.
}
\examples{
x <- runif(100)
y <- x + rnorm(100)
z <- x + max(y)
toyData <- cbind(x,y,z)
### Finish examples
}

