% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.r
\name{transforms}
\alias{transforms}
\title{Transform data in various ways for easy downstream manipulation.}
\usage{
transforms(x, type = c("zscore"), ...)
}
\arguments{
\item{x}{A matrix, data.frame, or data.table with the data of
interest.}

\item{type}{A character list of the transformations desired. See
'Details'}
}
\value{
A list of the transformed data as data.tables.
}
\description{
Transform data in various ways for easy downstream manipulation.
}
\details{
If there are negatives present, using any of log transforms
will generate an error. Types of transformations incldue "Zscore" -- 
transforms to z-scores, "rank" -- passes to ranks, "1e3" -- scales 
the data between $[0,1e3]$, "log10" -- checks for negatives
and then removes 0's and takes the $\log_10$ of the data, "log" --
same as "log10" but for base $e$.  "slog10" -- checks for negatives
and then removes 0's and takes $\log_10$, then zscores. "slog" --
same as "slog10" but for base $e$. And "all" performs all of the
above.
}
\examples{
x <- runif(100)
y <- x + rnorm(100)
z <- x + max(y)
toyData <- cbind(x,y,z)
transData <- transforms(toyData,type=c("all"),
ties.method='average')

}

